Split lambda endpoints into separate files.
